+++ src/main/java/calculator/model/implement/DelimiterChecker.java
@@ -1,4 +1,4 @@
-package calculator.model;
+package calculator.model.implement;
 
 import static calculator.common.DelimiterConstant.CUSTOM_DELIMITER_NOT_FOUND;
 import static calculator.common.DelimiterConstant.DELIMITER_CREATOR_BACK;
