+++ src/main/java/calculator/model/implement/DelimiterManager.java
@@ -1,4 +1,4 @@
-package calculator.model;
+package calculator.model.implement;
 
 import java.util.List;
