+++ src/main/java/calculator/model/implement/DelimiterStore.java
@@ -1,5 +1,6 @@
-package calculator.model;
+package calculator.model.implement;
 
+import calculator.model.domain.DefaultDelimiter;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
