+++ src/test/java/calculator/model/DelimiterManagerTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 import calculator.fixture.CalculatorAppFixture;
+import calculator.model.implement.DelimiterManager;
+import calculator.model.implement.DelimiterStore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
