+++ src/test/java/calculator/model/StringSumCalculatorTest.java
@@ -2,9 +2,9 @@ package calculator.model;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import calculator.fixture.CalculatorAppFixture;
+import calculator.model.service.StringSumCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
